CNN for CIFAR-10/requirements.txt
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
docopt==0.6.2
enum34==1.1.6
graphviz==0.8.1
h5py==2.7.1
html5lib==0.9999999
jedi==0.11.1
Keras==2.1.2
keras-vis==0.4.1
Markdown==2.6.10
matplotlib==2.1.1
networkx==2.0
numpy==1.13.3
olefile==0.44
parso==0.1.1
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptpython==0.41
pydot==1.2.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
wcwidth==0.1.7
Werkzeug==0.13
